+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Markdown-style Docstring support for global definitions in user library.
+- Provide completion for documented variables in user library.
 
 ### Fixed